// src/engines/cunny/generated/reference-models.ts
export interface CuNNyGeneratedReferenceModelMeta {
  key: string;
  id: string;
  name: string;
  variant: string;
  sourceFile: string;
  directory: string;
  stageCount: number;
}

export const cunnyGeneratedReferenceModelMetas: CuNNyGeneratedReferenceModelMeta[] = [
  {
    "key": "CUNNY_2X12_DS",
    "id": "cunny/Upscale/DS/2x12",
    "name": "CuNNy 2x12 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-2x12-DS.glsl",
    "directory": "2x12_ds",
    "stageCount": 4
  },
  {
    "key": "CUNNY_3X12_DS",
    "id": "cunny/Upscale/DS/3x12",
    "name": "CuNNy 3x12 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-3x12-DS.glsl",
    "directory": "3x12_ds",
    "stageCount": 5
  },
  {
    "key": "CUNNY_4X12_DS",
    "id": "cunny/Upscale/DS/4x12",
    "name": "CuNNy 4x12 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-4x12-DS.glsl",
    "directory": "4x12_ds",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X16_DS",
    "id": "cunny/Upscale/DS/4x16",
    "name": "CuNNy 4x16 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-4x16-DS.glsl",
    "directory": "4x16_ds",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X24_DS",
    "id": "cunny/Upscale/DS/4x24",
    "name": "CuNNy 4x24 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-4x24-DS.glsl",
    "directory": "4x24_ds",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X32_DS",
    "id": "cunny/Upscale/DS/4x32",
    "name": "CuNNy 4x32 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-4x32-DS.glsl",
    "directory": "4x32_ds",
    "stageCount": 6
  },
  {
    "key": "CUNNY_8X32_DS",
    "id": "cunny/Upscale/DS/8x32",
    "name": "CuNNy 8x32 DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-8x32-DS.glsl",
    "directory": "8x32_ds",
    "stageCount": 10
  },
  {
    "key": "CUNNY_FAST_DS",
    "id": "cunny/Upscale/DS/Fast",
    "name": "CuNNy fast DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-fast-DS.glsl",
    "directory": "fast_ds",
    "stageCount": 4
  },
  {
    "key": "CUNNY_FASTER_DS",
    "id": "cunny/Upscale/DS/Faster",
    "name": "CuNNy faster DS",
    "variant": "ds",
    "sourceFile": ".reference/CuNNy/mpv/ds/CuNNy-faster-DS.glsl",
    "directory": "faster_ds",
    "stageCount": 4
  },
  {
    "key": "CUNNY_2X12_SOFT",
    "id": "cunny/Upscale/SOFT/2x12",
    "name": "CuNNy 2x12 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-2x12-SOFT.glsl",
    "directory": "2x12_soft",
    "stageCount": 4
  },
  {
    "key": "CUNNY_3X12_SOFT",
    "id": "cunny/Upscale/SOFT/3x12",
    "name": "CuNNy 3x12 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-3x12-SOFT.glsl",
    "directory": "3x12_soft",
    "stageCount": 5
  },
  {
    "key": "CUNNY_4X12_SOFT",
    "id": "cunny/Upscale/SOFT/4x12",
    "name": "CuNNy 4x12 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-4x12-SOFT.glsl",
    "directory": "4x12_soft",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X16_SOFT",
    "id": "cunny/Upscale/SOFT/4x16",
    "name": "CuNNy 4x16 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-4x16-SOFT.glsl",
    "directory": "4x16_soft",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X24_SOFT",
    "id": "cunny/Upscale/SOFT/4x24",
    "name": "CuNNy 4x24 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-4x24-SOFT.glsl",
    "directory": "4x24_soft",
    "stageCount": 6
  },
  {
    "key": "CUNNY_4X32_SOFT",
    "id": "cunny/Upscale/SOFT/4x32",
    "name": "CuNNy 4x32 SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-4x32-SOFT.glsl",
    "directory": "4x32_soft",
    "stageCount": 6
  },
  {
    "key": "CUNNY_FAST_SOFT",
    "id": "cunny/Upscale/SOFT/Fast",
    "name": "CuNNy fast SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-fast-SOFT.glsl",
    "directory": "fast_soft",
    "stageCount": 4
  },
  {
    "key": "CUNNY_FASTER_SOFT",
    "id": "cunny/Upscale/SOFT/Faster",
    "name": "CuNNy faster SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-faster-SOFT.glsl",
    "directory": "faster_soft",
    "stageCount": 4
  },
  {
    "key": "CUNNY_VERYFAST_SOFT",
    "id": "cunny/Upscale/SOFT/Veryfast",
    "name": "CuNNy veryfast SOFT",
    "variant": "soft",
    "sourceFile": ".reference/CuNNy/mpv/soft/CuNNy-veryfast-SOFT.glsl",
    "directory": "veryfast_soft",
    "stageCount": 4
  }
];